requirements.txt
discord.py==1.7.2
sentry-sdk==1.1.0
tortoise-orm==0.17.2
asyncpg==0.22.0
python-dotenv==0.17.1
Django==3.2.2
django-grappelli==2.15.1
psycopg2-binary==2.8.6
django-environ==0.4.5
